requirements.txt
fastapi
uvicorn
ocpp
python-dotenv
websockets
pytest-asyncio
pytest
httpx
vcrpy